Split up `non_copy_const` ui test
